Ensure pluck, count, first, etc work
Implements ability to reuse scope for pluck, first, count, etc and
updates the tests.

This implementation requires no changes to Rails 6.0 or Rails 5.2.

In order to ensure pluck and friends work correctly we need to not map
all the join ids at the end of the scope chain loop and instead select
the ids from the last reflection in the chain.

Co-authored-by: Aaron Patterson <aaron.patterson@gmail.com>