Failing test for Polymorphic Relation
This test fails a polymorphic relation that is represented by the
Favoritable object. An Employee should be able to favorite either a Dock
or a Ship.

The test currently fails with the following error

```
NoMethodError: undefined method `map!' for #<ActiveRecord::Relation []>
```

Upon inspection this returns a `CollectionProxy` with the correct data.
Maybe we just need to return the correct type of object from the
overriden `scope` method